Split DpmService into multiple files according to resource type
