+++ services/data_plane_manager/src/main/java/com/futurewei/alcor/dataplane/service/ovs/NeighborService.java
@@ -1,0 +1,131 @@
+/*
+Copyright 2019 The Alcor Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+        you may not use this file except in compliance with the License.
+        You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+        Unless required by applicable law or agreed to in writing, software
+        distributed under the License is distributed on an "AS IS" BASIS,
+        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+        See the License for the specific language governing permissions and
+        limitations under the License.
+*/
+package com.futurewei.alcor.dataplane.service.ovs;
+
+import com.futurewei.alcor.dataplane.entity.MulticastGoalState;
+import com.futurewei.alcor.dataplane.entity.UnicastGoalState;
+import com.futurewei.alcor.dataplane.exception.NeighborInfoNotFound;
+import com.futurewei.alcor.dataplane.exception.PortFixedIpNotFound;
+import com.futurewei.alcor.schema.Common;
+import com.futurewei.alcor.schema.Neighbor;
+import com.futurewei.alcor.schema.Port;
+import com.futurewei.alcor.web.entity.dataplane.NeighborEntry;
+import com.futurewei.alcor.web.entity.dataplane.NeighborInfo;
+import com.futurewei.alcor.web.entity.dataplane.v2.NetworkConfiguration;
+import org.springframework.stereotype.Service;
+
+import java.util.*;
+
+@Service
+public class NeighborService extends ResourceService {
+    public Neighbor.NeighborState buildNeighborState(NeighborEntry neighborEntry, NeighborInfo neighborInfo, Common.OperationType operationType) {
+        Neighbor.NeighborConfiguration.Builder neighborConfigBuilder = Neighbor.NeighborConfiguration.newBuilder();
+        neighborConfigBuilder.setRevisionNumber(FORMAT_REVISION_NUMBER);
+        //neighborConfigBuilder.setId(); // TODO: We are going to need this per latest ACA change
+        neighborConfigBuilder.setVpcId(neighborInfo.getVpcId());
+        //neighborConfigBuilder.setName();
+        neighborConfigBuilder.setMacAddress(neighborInfo.getPortMac());
+        neighborConfigBuilder.setHostIpAddress(neighborInfo.getHostIp());
+        Neighbor.NeighborType neighborType = Neighbor.NeighborType.valueOf(neighborEntry.getNeighborType().getType());
+
+        //TODO:setNeighborHostDvrMac
+        //neighborConfigBuilder.setNeighborHostDvrMac();
+        Neighbor.NeighborConfiguration.FixedIp.Builder fixedIpBuilder = Neighbor.NeighborConfiguration.FixedIp.newBuilder();
+        fixedIpBuilder.setSubnetId(neighborInfo.getSubnetId());
+        fixedIpBuilder.setIpAddress(neighborInfo.getPortIp());
+        fixedIpBuilder.setNeighborType(neighborType);
+        neighborConfigBuilder.addFixedIps(fixedIpBuilder.build());
+        //TODO:setAllowAddressPairs
+        //neighborConfigBuilder.setAllowAddressPairs();
+
+        Neighbor.NeighborState.Builder neighborStateBuilder = Neighbor.NeighborState.newBuilder();
+        neighborStateBuilder.setOperationType(operationType);
+        neighborStateBuilder.setConfiguration(neighborConfigBuilder.build());
+
+        return neighborStateBuilder.build();
+    }
+
+    public void buildNeighborStates(NetworkConfiguration networkConfig, String hostIp,
+                                     UnicastGoalState unicastGoalState,
+                                     MulticastGoalState multicastGoalState) throws Exception {
+        Map<String, NeighborInfo> neighborInfos = networkConfig.getNeighborInfos();
+        if (neighborInfos == null || neighborInfos.size() == 0) {
+            return;
+        }
+
+        Map<String, List<NeighborEntry>> neighborTable = networkConfig.getNeighborTable();
+        if (neighborTable == null || neighborTable.size() == 0) {
+            return;
+        }
+
+        List<Port.PortState> portStates = unicastGoalState.getGoalStateBuilder().getPortStatesList();
+        if (portStates == null || portStates.size() == 0) {
+            return;
+        }
+
+        List<NeighborEntry> multicastNeighborEntries = new ArrayList<>();
+        for (Port.PortState portState: portStates) {
+            List<Port.PortConfiguration.FixedIp> fixedIps = portState.getConfiguration().getFixedIpsList();
+            if (fixedIps == null) {
+                throw new PortFixedIpNotFound();
+            }
+
+            for (Port.PortConfiguration.FixedIp fixedIp: fixedIps) {
+                List<NeighborEntry> neighborEntries = neighborTable.get(fixedIp.getIpAddress());
+                if (neighborEntries == null) {
+                    throw new NeighborInfoNotFound();
+                }
+
+                for (NeighborEntry neighborEntry: neighborEntries) {
+                    NeighborInfo neighborInfo = neighborInfos.get(neighborEntry.getNeighborIp());
+                    if (neighborInfo == null) {
+                        throw new NeighborInfoNotFound();
+                    }
+
+                    if (hostIp.equals(neighborInfo.getHostIp())) {
+                        continue;
+                    }
+
+                    unicastGoalState.getGoalStateBuilder().addNeighborStates(buildNeighborState(
+                            neighborEntry, neighborInfo, networkConfig.getOpType()));
+                }
+
+                multicastNeighborEntries.addAll(neighborEntries);
+            }
+        }
+
+        Set<NeighborInfo> neighborInfoSet = new HashSet<>();
+        for (NeighborEntry neighborEntry: multicastNeighborEntries) {
+            String localIp = neighborEntry.getLocalIp();
+            String neighborIp = neighborEntry.getNeighborIp();
+            NeighborInfo neighborInfo1 = neighborInfos.get(localIp);
+            NeighborInfo neighborInfo2 = neighborInfos.get(neighborIp);
+            if (neighborInfo1 == null || neighborInfo2 == null) {
+                throw new NeighborInfoNotFound();
+            }
+
+            if (!multicastGoalState.getHostIps().contains(neighborInfo2.getHostIp())) {
+                multicastGoalState.getHostIps().add(neighborInfo2.getHostIp());
+            }
+
+            if (!neighborInfoSet.contains(neighborInfo1)) {
+                multicastGoalState.getGoalStateBuilder().addNeighborStates(buildNeighborState(
+                        neighborEntry, neighborInfo1, networkConfig.getOpType()));
+                neighborInfoSet.add(neighborInfo1);
+            }
+        }
+    }
+}
