+++ services/data_plane_manager/src/main/java/com/futurewei/alcor/dataplane/service/ovs/ResourceService.java
@@ -1,0 +1,37 @@
+/*
+Copyright 2019 The Alcor Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+        you may not use this file except in compliance with the License.
+        You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+        Unless required by applicable law or agreed to in writing, software
+        distributed under the License is distributed on an "AS IS" BASIS,
+        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+        See the License for the specific language governing permissions and
+        limitations under the License.
+*/
+package com.futurewei.alcor.dataplane.service.ovs;
+
+import com.futurewei.alcor.schema.Common.OperationType;
+
+public class ResourceService {
+    protected static final int FORMAT_REVISION_NUMBER = 1;
+
+    protected OperationType getOperationType(com.futurewei.alcor.common.enumClass.OperationType operationType) {
+        switch (operationType) {
+            case CREATE:
+                return OperationType.CREATE;
+            case UPDATE:
+                return OperationType.UPDATE;
+            case INFO:
+                return OperationType.INFO;
+            case DELETE:
+                return OperationType.DELETE;
+            default:
+                return OperationType.UNRECOGNIZED;
+        }
+    }
+}
